+++ server/src/handlers/validation.js
@@ -1,0 +1,9 @@
+const {validationResult} = require('express-validator')
+
+exports.validate = (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(400).json({errors: errors.array()});
+  }
+  next()
+};
